requirements.txt
streamlit
pandas
requests
scikit-learn
matplotlib
seaborn
python-dotenv
geopy
pytz
folium